% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb-package.R
\docType{package}
\name{gedsb-package}
\alias{gedsb}
\alias{gedsb-package}
\title{gedsb: GEDSB Visual Identity}
\description{
This is a package for custom visual identity for the Grand Erie District School Board.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grousell/gedsb}
  \item Report bugs at \url{https://github.com/grousell/gedsb/issues}
}

}
\author{
\strong{Maintainer}: Greg Rousell \email{grousell@gmail.com}

}
\keyword{internal}
