% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_colours.R
\name{gedsb_colours}
\alias{gedsb_colours}
\title{gedsb_colours}
\usage{
gedsb_colours(...)
}
\arguments{
\item{cols}{}
}
\value{
Hex colour for GEDSB visual identity
}
\description{
gedsb_colours
}
\examples{
gedsb_colours("Blue")
}
