% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_colours_fonts.R
\name{gedsb_colours_fonts}
\alias{gedsb_colours_fonts}
\title{gedsb_colours}
\usage{
gedsb_colours_fonts(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
R objects with names Hex colours for GEDSB visual identity
}
\description{
gedsb_colours
}
\examples{
# gedsb_colours_fonts()

}
