% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_ggplot_theme_flipped_bar_chart.R
\name{gedsb_flipped_bar_plot_theme}
\alias{gedsb_flipped_bar_plot_theme}
\title{GEDSB ggplot2 Theme for Flipped Bar Charts}
\usage{
gedsb_flipped_bar_plot_theme()
}
\value{
Sets gglot2 theme for flipped bar charts
}
\description{
GEDSB ggplot2 Theme for Flipped Bar Charts
}
\examples{
gedsb_flipped_bar_plot_theme()
}
