% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_get_survey_data.R
\name{gedsb_get_survey_data}
\alias{gedsb_get_survey_data}
\title{Get Survey Data
A function to download data from Snap Surveys}
\usage{
gedsb_get_survey_data(survey_name, url)
}
\arguments{
\item{survey_name}{The name of survey to be used, preferable snake case (e.g. my_survey)}

\item{url}{URL from the Analyse tab on Snap Surveys.}
}
\value{
.txt file, and two .csv files. One for survey data, one with a Data Dictionary
}
\description{
Get Survey Data
A function to download data from Snap Surveys
}
\examples{
# gedsb_get_survey_data(survey_name, url)

}
