% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_ggplot_theme.R
\name{gedsb_ggplot_theme}
\alias{gedsb_ggplot_theme}
\title{GEDSB ggplot2 Theme for Bar Charts}
\usage{
gedsb_ggplot_theme()
}
\value{
Sets gglot2 theme
}
\description{
GEDSB ggplot2 Theme for Bar Charts
}
\examples{
gedsb_bar_plot_theme()
}
