% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_import_font.R
\name{gedsb_import_font}
\alias{gedsb_import_font}
\title{gedsb_import_fonts}
\usage{
gedsb_import_font(...)
}
\arguments{
\item{...}{}
}
\value{
Loads Windows Fonts
}
\description{
gedsb_import_fonts
}
\examples{
gedsb_import_font()
}
