% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_factor_levels.R
\name{gedsb_likert_factor_levels}
\alias{gedsb_likert_factor_levels}
\title{Likert Factor Levels}
\usage{
gedsb_likert_factor_levels(x)
}
\arguments{
\item{x}{}
}
\value{
Vector of factor, with levels "Strongly Disagree", "Disagree", "Neither Agree or Disagree", "Agree", "Strongly Agree"
}
\description{
Likert Factor Levels
}
\examples{
gedsb_likert_factor_levels()

}
