% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_report_template_pdf.R
\name{gedsb_report_template_pdf}
\alias{gedsb_report_template_pdf}
\title{Report Temnplate - PDF
A function that downloads a .qmd template, images and .tex files for a Grand Erie report template.}
\usage{
gedsb_report_template_pdf(report_name)
}
\arguments{
\item{report_name}{Name of the report}
}
\description{
Report Temnplate - PDF
A function that downloads a .qmd template, images and .tex files for a Grand Erie report template.
}
\examples{
# gedsb_report_template_pdf(test_report)
}
