% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_significant_factor_levels.R
\name{gedsb_significant_factor_levels}
\alias{gedsb_significant_factor_levels}
\title{Significant Factor Levels
Converts a character vector of "1 - Not at all",  "2", "3", "4", "5 - Significantly" to ordered factors.}
\usage{
gedsb_significant_factor_levels(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Vector of factors with levels "1 - Not at all",  "2", "3", "4", "5 - Significantly" to ordered factors.
}
\description{
Significant Factor Levels
Converts a character vector of "1 - Not at all",  "2", "3", "4", "5 - Significantly" to ordered factors.
}
\examples{
# gedsb_significant_factor_levels(x)

}
