% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_summary_table.R
\name{gedsb_summary_table}
\alias{gedsb_summary_table}
\title{GEDSB Summary Table}
\usage{
gedsb_summary_table(df, group_1, group_2, ...)
}
\arguments{
\item{df}{A dataframe}

\item{group_1}{The first grouping variable}

\item{group_2}{The second grouping variable (optional)}

\item{...}{Addtional arguments}
}
\value{
\link[tibble:tibble-package]{A tibble} with columns for \code{group_1}, \code{group_2},
\code{n} and \code{perc}
}
\description{
This function will create a one or two variable summary table with percentages
}
\examples{
# dat <- palmerpenguins::penguins
# gedsb_summary_table (dat, species)
# gedsb_summary_table (dat, species, sex)

}
