% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_understanding_factor_level.R
\name{gedsb_understanding_factor_levels}
\alias{gedsb_understanding_factor_levels}
\title{Understanding Factor Levels
Converts a character vector of "1 - None",  "2", "3", "4", "5 - A Great Deal" to ordered factors.}
\usage{
gedsb_understanding_factor_levels(x)
}
\arguments{
\item{x}{}
}
\value{
Vector of factors with levels "1 - None",  "2", "3", "4", "5 - A Great Deal" to ordered factors.
}
\description{
Understanding Factor Levels
Converts a character vector of "1 - None",  "2", "3", "4", "5 - A Great Deal" to ordered factors.
}
\examples{
gedsb_understanding_factor_levels(x)
}
