% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedsb_likert_factor_levels.R
\name{gedsb_likert_factor_levels}
\alias{gedsb_likert_factor_levels}
\title{Likert Factor Levels
Converts a character vector of "Strongly Disagree",  "Disagree", "Neither Agree or Disagree", "Agree", "Strongly Agree" to ordered factors.}
\usage{
gedsb_likert_factor_levels(x)
}
\arguments{
\item{x}{}
}
\value{
Vector of factors with levels "Strongly Disagree",  "Disagree", "Neither Agree or Disagree", "Agree", "Strongly Agree"
}
\description{
Likert Factor Levels
Converts a character vector of "Strongly Disagree",  "Disagree", "Neither Agree or Disagree", "Agree", "Strongly Agree" to ordered factors.
}
\examples{
test <- c(rep ("Strongly Disagree", 2),
rep ("Disagree", 3),
rep ("Neither Agree or Disagree", 2),
rep ("Agree", 1),
rep ("Strongly Agree", 3))

gedsb_likert_factor_levels(test)
}
